IGNITE-25748 Address code duplication
